Merge am_endpoint_handler() and am_handler().
We currently handle requests to many endpoints from the access control
hooks. This change bypasses access control in those cases, and handles
the requests from the "handler" hook instead.

This change is necessary to be able to do anything else than redirects
from the handlers. As a side effect, it also simplifies the code.

git-svn-id: https://modmellon.googlecode.com/svn/trunk@144 a716ebb1-153a-0410-b759-cfb97c6a1b53