Do not mark a session as logged in before all processing is completed.
If we mark the session as logged in before all processing is completed,
a failure during login processing (e.g. too big attribute values) can
cause a user to receive a "half-authenticated" session.

This patch changes the code so that the last task before releaseing the
session is to mark it as logged in.

git-svn-id: https://modmellon.googlecode.com/svn/trunk@184 a716ebb1-153a-0410-b759-cfb97c6a1b53