Fix MellonDoNotVerifyLogoutSignature.
The configuration directive declaration was missing the parameter
describing where the data from the option was stored. The result is
that we access invalid memory during configuration parsing, leading to
a segmentation fault.

As far as I can tell, this error has always been present, so this
option hasn't worked before.

git-svn-id: https://modmellon.googlecode.com/svn/trunk@220 a716ebb1-153a-0410-b759-cfb97c6a1b53