Fix memory leak of loaded IdPs.
We free the first element of a linked list, but not the data nor the
subsequent elements. Fix that by first iterating through the list
freeing the elements and then freeing the list using g_list_free().

We could have used g_list_free_full, but that requires version 2.28
of GLib, which is still slightly too recent.

git-svn-id: https://modmellon.googlecode.com/svn/trunk@237 a716ebb1-153a-0410-b759-cfb97c6a1b53