Use "long" type in curl_easy_setopt for timeouts
Curl timeouts in auth_mellon_httpclient should be given as long
values and have semantic meaning of seconds.

The code currently passes them in as apr_time_t type which
contains seconds, although apr_time_t should contain
microseconds.

I suggest to not use apr_time_t here, because it is misleading,
and instead use a plain int. The code calling httpclient already
prepares the value as an int. Furthermore convert the value to a
long when curl_easy_setopt() is being called.