Fix segmentation fault with POST field without a value.
This patch fixes a segmentation fault that can occur after the user
has logged in if the server is configured to replay POST data from
before login. If the POST data contained a field without a value we
would attempt to pass a constant string to the
am_urldecode()-function, which would crash with a segmentation fault.

This patch fixes that by using an empty string allocated on the stack
instead of a constant string.

Fixes #115.