Make MellonUser case-insensitive.
MellonUser used to perform its attribute match in a case-sesnsitive
manner, while environment variables are stored in a case-insensitive
table in Apache.

The result is a bit of inconsistency between the way mod_auth_mellon
handles the attribute names and the way they are accessed other places
in Apache.

This patch changes the code to use a case-insensitive match when
processing the MellonUser directive.

Fixes issue #131.