Add user_guide to distribution, use AC_DEFINE instead of CFLAGS
This patch corrects a few minor autotool issues.

The user_guide was recently added but that commit failed to include
adding the new documentation to the tarball, Makefile.in was augmented
to include the new files to the list of distribution files.

Formerly the #defines generated by configure were passed to the
compiler on the command line in various CFLAGS values. Although that
works the more -Dxxx that configure generates the longer the compile
command becomes and it starts to get unreadable and possibly exceed
command line length. A more common practice with autotools is to
employ autoheader whereby configure generates a file typically called
config.h which then is included by the C files. The contents of
config.h contains the #defines as generated by configure. configure.ac
was updated to utilize the AC_DEFINE in lieu of adding -Dxxx to CFLAGS
and to generate and output config.h.

Note: autogen.sh needs to be re-run to pick up these changes so that
the configure included in the tarball contains the updated version.

Signed-off-by: John Dennis <jdennis@redhat.com>