Add NameID discussion to User Guide
How NameIDs are used in SAML often confuse people trying to configure
a service provider especially when they try to emulate a userid. This
patch adds several sections describing the concept of a SAML NameID,
how the use of NameIDs are configured in SAML and some suggested
approaches to utilizing NameIDs.

Signed-off-by: John Dennis <jdennis@redhat.com>