Connect lasso log messages to mellon's logging
Mellon utilizes the Lasso library for most of it's SAML
operations. Therefore access to Lasso's log messages are often useful
when diagnosing SAML issues. This patch adds a logging callback
handler that is passed into Lasso. When Lasso emits a message the
callback handler is invoked. The Lasso log level is translated to a
Apache log level, tagged with Lasso's log domain and then emitted as a
message from the mellon module.

Signed-off-by: John Dennis <jdennis@redhat.com>